Enable spellchecking
Create 'spell' target that uses pospell tool to run spell checking
on the translation files. The GitHub workflow will create an
artifact containing log files, making it possible to donwload and
study possible typos.

The 'dict' file is a custom dictionary with words known to be
found in the translation files, but that should not be considered
as typo.