chore: Revamp our CI solution
This commit reimplements the CI solution for python-docs-pt-br
improving several aspects over the previous solution.

PREVIOUS SOLUTION

The previous solution had a Makefile coded with all operations,
and a single workflow (pythonpackage.yml) that had used this
Makefile. The problem with this implementation was:

 - The Makefile had operations that would never be used in the
   user command line, hence they did not have to be in a script
   for the CLI causing the script to be bloated
 - The Makefile syntax is not very easy to read for everyone,
   while the shell script syntax is more common.
 - The workflow was bloated with different tasks, hence making
   it impossible to best use the events the could trigger the
   jobs (update, build, etc.)
 - The workflow could not be triggered manually by the admin
 - The spellchecking output generated by the target in Makefile
   and the job in the workflow had to much noise making it hard
   to read, and no one cared about to.

Also, some workflow instructions could receive some improvements.
For instance, the Telegram notification message just linked to
the workflow run without any other useful information on the
error report. Instead of requiring the user to go to a link,
it is more efficient to provide the error message already.

NEW SOLUTION

The new CI solution drops the use of the Makefile and the single
workflow file to implement a few shell scripts and workflow files.

It features the following workflows:
 - update: pull translations from Transifex, and push into the
           main branch and also previous branches
 - build: build translated docs and notify in case of errors
 - compendium: generate a PO compendium of all translated strings

The build workflow makes use of a problem matcher to annotate
Sphinx errors, making them more visible.

The workflows use the following shell scripts:
 - build: call sphinx-build to build the documentation
 - prepmsg: prepare a message to notify in a Telegram target group
 - update: call POT update, tx pull and other update instructions

The update workflow is scheduled to run every night, and will call
the build and compendium workflow -- so all actions are run on
schedule event.

These workflows can also be run manually in the webUI by the
repository admin thanks to workflow_dispatch event, which is
useful e.g. a hackathon/sprint/etc. where one wants to run the
workflow more than one time (scheduled) a day.

The command for pulling translations from Transifex, tx pull, is
called by update.sh without --force but with --use-git-timestamps.
This allows us to make sure all translations are pulled correct,
without the 40 minutes of forced download.

The build.sh can be run in the command-line interface. All it is
required is to clone cpython repo in the python-docs-pt-br root
directory and to install the requirements from 'requirements.txt'
and 'cpython/Doc/requirement.txt' files.

build.sh uses pt_BR language code by default, but can be used by
another language teams by passing other languages.
e.g. 'build.sh fr'

SETTINGS NEEDED IN THE REPOSITORY

- New TX_TOKEN secret required

The previous workflow using TRANSIFEX_APIKEY as the Transifex API
token variable name. As per Transifex documentation, when TX_TOKEN
it is not required to create configuration file, hence it is much
simpler in CI environment to use TX_TOKEN than other custom names.

TX_TOKEN has to be created with in order to the update workflow to
works correctly.

- Disable the previous workflow

The .github/workflows/pythonpackage.yml, named "Build and update
documentation in the Actions page, is scheduled to run every night,
just like the new CI solution. So, there's a conflict.

This workflow must be disabled. Please refer the GitHub docs on how
to disable an workflow:
https://docs.github.com/en/actions/managing-workflow-runs/disabling-and-enabling-a-workflow